court-ai-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
supabase==2.18.1
email-validator==2.2.0
pandas==2.3.2
numpy==2.2.6
scikit-learn==1.7.2
joblib==1.4.2
